Move permissions selectors to existing hook
